star tasks 1 and 2, weekly next date support and refactor
