+++ .env
@@ -1,0 +1,2 @@
+TODO_PORT="7540"
+TODO_DBFILE="scheduler.db"
