+++ tests/settings.go
@@ -2,6 +2,6 @@ package tests
 
 var Port = 7540
 var DBFile = "../scheduler.db"
-var FullNextDate = false
+var FullNextDate = true
 var Search = true
 var Token = ``
